+++ client/src/App.tsx
@@ -4,7 +4,7 @@ import { ThemeProvider } from 'styled-components';
 
 import theme from '@style/theme';
 import GlobalStyle from '@style/GlobalStyle';
-import { FanUP, Home, Login, NotFound, AuthCallback } from './Routes';
+import { FanUP, Home, Login, NotFound, AuthCallback, Tickets } from './Routes';
 
 function App() {
     return (
@@ -15,7 +15,8 @@ function App() {
                     <Route path="/login" element={<Login />} />
                     <Route path="/auth/:domain/callback" element={<AuthCallback />} />
                     <Route path="/" element={<Home />} />
-                    <Route path="/FanUP/:fanUpId" element={<FanUP />} />
+                    <Route path="/fanUp/:fanUpId" element={<FanUP />} />
+                    <Route path="/tickets" element={<Tickets />} />
                     <Route path="*" element={<NotFound />} />
                 </Routes>
             </Suspense>
