+++ client/src/components/UI/atoms/CameraBtn/index.tsx
