+++ client/src/components/UI/atoms/CameraBtn/test/index.test.tsx
