+++ client/src/components/UI/atoms/ExitBtn/index.tsx
