+++ client/src/components/UI/atoms/MuteBtn/index.tsx
