+++ client/src/components/UI/atoms/Participants/index.tsx
