+++ client/src/pages/Tickets/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Tickets = () => {
+    return <div>Tickets</div>;
+};
+
+export default Tickets;
